+++ app/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod conn;
-pub mod fetcher;
 pub mod migration;
 pub mod modifier;
 mod utils;
+pub mod views;
