+++ app/src/views/activity_view.rs
