+++ app/src/views/search_view.rs
