+++ app/src/views/tx_view.rs
